Phase 1 - Employee Class, readCSV Method Wrap Up & Testing
Created a class to represent an Employee object. The employee object contains variables that represent each column in a employees record (first name, salary etc.). Expanded further on the readCSV method and made the data that is read in be stored as an Employee and then in a HashMap using the employeeID as the key. Made it so the readCSV method excludes duplicates ready to catch them and store them elsewhere. Whilst developing this method further a TDD approach was used by developing tests to ensure exceptions are caught and some simple behavior's are tested. Further tests to be writen. Also made sure to implement two new test csv's for JUnit testing.