Phase 2 - DataVerification Class Bugfixes & Wrap Up
Fixed bug with DataVerification class whereby it was not checking the Employee objects properly due to the nature of the checks. Cleaned up the class and commented it out.  Cleaned up FileReader class and removed any redundant/unwanted code.  Fixed a bug with FileWriter class where records were being written over on the first line.