Phase 3 - Implementing Threading & Code Cleanup
Added a separate static method within the FileReader class which can take in a HashMap, Employee and employeeID and check that that there are no duplicate or corrupt pieces of data. Added in the "synchronized" keyword to the FileWriter and FileReader class so that they can handle multithreading correctly. Cleaned up the Thread class. Commented multiple classes. Altered the overridden toString method in Employee class for a cleaner looking output.