Phase 3 - Log4J Implemented & CSV Folders Made
Implemented the Log4j dependency into the pom.xml file. Imported and instantiated the logger into the ThreadFileReader class and tested that it outputs to the csvmigrationlogfile.log in the resources folder. Also moved the csv files into folders called "csvInputs" or "csvOutputs" depending of their use. Also updated filepath for CSV file writing to the "csvOutputs" folder.